refactor(ligas): simplify RankingsPage PR — merge tests, remove unused files
Merge states and interactions tests into single RankingsPage.test.tsx (5 tests).
Remove unused rankingConfig.ts, leagueService.test.ts and index.css variables.
Simplify App.test.tsx route test.
