--- frontend/src/config/rankingConfig.ts
@@ -1,9 +1,0 @@
-export const TOP_RANGE = {
-  min: 1,
-  max: 3,
-};
-
-export const DANGER_RANGE = {
-  min: 8,
-  max: 10,
-};
