+++ frontend/src/index.css
@@ -16,8 +16,6 @@
 
   --color-primary: #b91879;
   --color-gray-foreground: #7e7e7e;
-  --highlight-top: #d1fae5;
-  --highlight-danger: #ffe4e6;
 }
 
 @theme {
