--- frontend/src/pages/__test__/RankingsPage.states.test.tsx
@@ -1,114 +1,0 @@
-import { render, screen, waitFor } from "@testing-library/react";
-import "@testing-library/jest-dom/vitest";
-import { describe, it, expect, vi, beforeEach } from "vitest";
-import RankingsPage from "../RankingsPage";
-import { getLeagueRanking } from "../../services/leagueService";
-
-vi.mock("../../services/leagueService", () => ({
-  getLeagueRanking: vi.fn(),
-}));
-
-vi.mock("../../components/LeagueToggle/LeagueToggle", () => ({
-  default: ({ view, onChange }: { view: string; onChange: (v: string) => void }) => (
-    <div>
-      <button onClick={() => onChange("weekly")} aria-pressed={view === "weekly"}>Lliga setmanal</button>
-      <button onClick={() => onChange("global")} aria-pressed={view === "global"}>Classificació general</button>
-    </div>
-  ),
-}));
-
-vi.mock("../../components/leagues/StandingsTable", () => ({
-  StandingsTable: () => (
-    <table>
-      <thead><tr><th>Posició</th><th>Usuari</th><th>Punts</th></tr></thead>
-    </table>
-  ),
-}));
-
-vi.mock("../../components/leagues/StandingsEmptyState", () => ({
-  StandingsEmptyState: () => <p>No hi han dades disponibles</p>,
-}));
-
-vi.mock("../../components/leagues/StandingsTableSkeleton", () => ({
-  StandingsTableSkeleton: () => <p>Carregant...</p>,
-}));
-
-const mockRanking = [
-  {
-    position: 1,
-    user_id: 101,
-    points: 94,
-    created_at: "2026-04-24T00:00:00Z",
-    updated_at: "2026-04-24T00:00:00Z",
-  },
-  {
-    position: 2,
-    user_id: 102,
-    points: 88,
-    created_at: "2026-04-24T00:00:00Z",
-    updated_at: "2026-04-24T00:00:00Z",
-  },
-];
-
-describe("RankingsPage — UI states", () => {
-  beforeEach(() => {
-    vi.clearAllMocks();
-  });
-
-  it("shows the skeleton while loading", () => {
-    vi.mocked(getLeagueRanking).mockReturnValue(new Promise(() => {}));
-
-    render(<RankingsPage />);
-
-    expect(screen.getByText("Carregant...")).toBeInTheDocument();
-  });
-
-  it("shows the toggle", async () => {
-    vi.mocked(getLeagueRanking).mockResolvedValue(mockRanking);
-
-    render(<RankingsPage />);
-
-    await waitFor(() => {
-      expect(screen.getByText("Classificació general")).toBeInTheDocument();
-      expect(screen.getByText("Lliga setmanal")).toBeInTheDocument();
-    });
-  });
-
-  it("shows the standings table after a successful fetch", async () => {
-    vi.mocked(getLeagueRanking).mockResolvedValue(mockRanking);
-
-    render(<RankingsPage />);
-
-    await waitFor(() => {
-      expect(screen.getByText("Posició")).toBeInTheDocument();
-      expect(screen.getByText("Usuari")).toBeInTheDocument();
-      expect(screen.getByText("Punts")).toBeInTheDocument();
-    });
-  });
-
-  it("shows the empty state when the API returns an empty array", async () => {
-    vi.mocked(getLeagueRanking).mockResolvedValue([]);
-
-    render(<RankingsPage />);
-
-    await waitFor(() => {
-      expect(
-        screen.getByText("No hi han dades disponibles"),
-      ).toBeInTheDocument();
-    });
-  });
-
-  it("shows an error message when the API fails", async () => {
-    vi.mocked(getLeagueRanking).mockRejectedValue(new Error("network error"));
-
-    render(<RankingsPage />);
-
-    await waitFor(() => {
-      expect(
-        screen.getByText(
-          "No s'han pogut carregar les dades. Torna-ho a intentar.",
-        ),
-      ).toBeInTheDocument();
-    });
-  });
-});
